Change dotenv-safe to dotenv
dotenv-safe was originally picked to assist in development, as it would
make sure you have your environment variables set correctly. However,
it poses an issue with the deployment environment, and unless we decide
to start messing around passing redundant environment variables to
containers, reverting to vanilla dotenv is the simplest option to get it
working.